Fixed broken link on Line 133 of README.md
I know this is super late and that this document might not be actively used, but I stumbled upon this repo while looking into Promises. I'm a current student at HR25 and doing the Oath micro-sprint :)

Only one change has been made to this document, detailed below:

On Line 133, 'Javascript Promises... in Wicket Detail' was linking to a fork of this exact same repo and not to the actual article itself. It has been replaced with the correct link.